# src/main/java/it/academy/cursebackspring/dto/request/CreateOrderDTO.java
package it.academy.cursebackspring.dto.request;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CreateOrderDTO {

    private Long userId;

    @NotNull
    private List<OrderItemDTO> orderItems;

}