# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: My REST API
  description: Some custom description of API.
  contact:
    name: Artem Yushkevich
    url: ''
    email: yushkevich@mail.ru
  license:
    name: License of API
    url: http://
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Dev server
security:
  - Bearer-Authentication: [ ]
paths:
...
  /api/v1/admin/update/product:
    put:
      description: Allows you to update an existing product
      tags:
        - Admin controller
      operationId: updateProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/update/order:
    put:
      description: Allows you to update the order status of an existing order
      tags:
        - Admin controller
      operationId: updateOrderStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/update/order/item:
    put:
      description: Allows you to update the item of an existing order
      tags:
        - Admin controller
      operationId: updateOrderItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/update/category:

    put:
      description: Allows you to update a existing category
      tags:
        - Admin controller
      operationId: updateCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/cart/add/item:

    post:
      description: Allows the user to add an item to their own shopping cart
      tags:
        - Cart controller
      operationId: addItemToCart
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddOrUpdateItemInCartDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/cart/update/item:

    post:
      description: Allows the user to update the item in their own shopping cart
      tags:
        - Cart controller
      operationId: addItemToCart_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddOrUpdateItemInCartDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/cart/add/order:

    post:
      description: Allows the user to create an order based on their shopping cart
      tags:
        - Cart controller
      operationId: addNewOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/admin/add/product:
    post:
      description: Allows you to add a new product
      tags:
        - Admin controller
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/add/order/item:
    post:
      description: Allows you to add a new order item to an existing order
      tags:
        - Admin controller
      operationId: addItemToOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/add/category:
    post:
      description: Allows you to add a new category
      tags:
        - Admin controller
      operationId: addCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonNode'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/admin/get/users:
    get:
      description: Allows you to obtain a list of all users
      tags:
        - Admin controller
      operationId: getUsers
      parameters:
        - name: dto
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/RequestDataDetailsDTO'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UsersDTO'
  /api/v1/admin/get/orders:
    get:
      description: Allows you to view all available orders
      tags:
        - Admin controller
      operationId: getOrders
      parameters:
        - name: dto
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/RequestDataDetailsDTO'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrdersDTO"
...
  /api/v1/cart/delete/item:
    delete:
      description: Allows you to remove the item from your shopping cart
      tags:
        - Cart controller
      operationId: deleteItemFromCart
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/delete/product/{productId}:
    delete:
      description: Allows you to delete a product by their ID
      tags:
        - Admin controller
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/get/order/{orderId}:
    delete:
      description: Allows you to get the items from the order
      tags:
        - Admin controller
      operationId: getOrderItems
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderItemsDTO'
  /api/v1/admin/delete/order/{id}:
    delete:
      description: Allows you to delete an order by their ID
      tags:
        - Admin controller
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/delete/order/item:
    delete:
      description: Allows you to delete an order item from an order
      tags:
        - Admin controller
      operationId: deleteOrderItem
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
        - name: orderId
          in: query
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/delete/category/{categoryId}/{root}:
    delete:
      description: This function allows you to delete a category by its ID. If this category is used by any products, you must use the root flag to delete the category with all products. However, if any of these products are used in orders, the root flag will not help you to delete the category.
      tags:
        - Admin controller
      operationId: deleteCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
        - name: root
          in: path
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
...
    UpdateProductDTO:
      description: New product data
      required:
        - categoryId
        - description
        - id
        - imageLink
        - name
        - price
      type: object
      properties:
        id:
          minimum: 1
          description: Targer product ID
          default: 1
          type: integer
          format: int64
        categoryId:
          default: 1
          description: Identifier of the associated category
          minimum: 1
          type: integer
          format: int64
        name:
          description: New name of the product
          default: "New product name"
          pattern: ^[A-zА-я0-9]{1,40}$
          type: string
        description:
          description: New product description
          type: string
        price:
          default: 2.2
          description: New product price
          minimum: 0.01
          exclusiveMinimum: false
          type: number
          format: double
        imageLink:
          default: /not-found-404.jpg
          description: New product image link
          type: string
    UpdateOrderStatusDTO:
      description: Updating the status of an existing order
      required:
        - orderId
        - orderStatus
      type: object
      properties:
        orderId:
          default: 1
          description: Target Order ID
          minimum: 1
          type: integer
          format: int64
        orderStatus:
          default: CANCELLED
          description: New order status
          type: string
          enum:
            - PENDING
            - PROCESSING
            - SHIPPED
            - DELIVERED
            - CANCELLED
    OrderItemDTO:
      description: Required data for adding a new or updating an existing order item
      required:
        - price
        - productId
        - quantity
      type: object
      properties:
        productId:
          default: 1
          description: Product ID
          minimum: 1
          type: integer
          format: int64
        orderId:
          default: 1
          description: Order ID
          type: integer
          format: int64
        quantity:
          default: 5
          description: Quantity of the item
          minimum: 1
          type: integer
          format: int64
        price:
          default: 5
          description: Total price for all items of this type
          minimum: 0.01
          exclusiveMinimum: false
          type: number
          format: double
    UpdateCategoryDTO:
      description: New category data
      required:
        - categoryId
        - categoryName
      type: object
      properties:
        categoryId:
          default: 1
          description: Target category ID
          minimum: 1
          type: integer
          format: int64
        categoryName:
          default: "New category name"
          description: New category name
          type: string
...
    AddOrUpdateItemInCartDTO:
      description: Required data for adding a new or updating an existing item in the cart
      required:
        - productId
        - quantity
      type: object
      properties:
        quantity:
          default: 5
          description: Quantity of the item
          minimum: 1
          type: integer
          format: int32
        productId:
          default: 1
          description: Target product ID
          minimum: 1
          type: integer
          format: int64
    CreateOrderDTO:
      description: Required data for creating a new order
      required:
        - orderItems
      type: object
      properties:
        orderItems:
          description: Included order items
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDTO'
...
    CreateProductDTO:
      description: Required data for creating a new product
      required:
        - categoryId
        - description
        - imageLink
        - name
        - price
      type: object
      properties:
        categoryId:
          description: Target category ID
          default: 1
          minimum: 1
          type: integer
          format: int64
        name:
          default: "ProductName"
          description: Product name
          pattern: ^[A-zА-я0-9]{1,40}$
          type: string
        description:
          default: "Product description"
          description: Product description
          type: string
        price:
          default: 2.2
          description: Product price
          minimum: 0.01
          exclusiveMinimum: false
          type: number
          format: double
        imageLink:
          default: "not-found-404.jpg"
          description: Product image link
          type: string
    JsonNode:
      type: object
    UserDTO:
      description: User data
      required:
        - id
        - name
        - phoneNumber
        - roles
        - surname
      type: object
      properties:
        id:
          description: User ID
          minimum: 1
          type: integer
          format: int64
        building:
          description: User building
          type: string
        city:
          description: User city
          type: string
        street:
          description: User street
          type: string
        email:
          description: User email
          type: string
        name:
          description: User name
          type: string
        phoneNumber:
          description: User phone number
          pattern: ^\+?(\d{1,3})?[-.\s]?\(?\d{1,4}\)?[-.\s]?\d{1,4}[-.\s]?\d{1,4}[-.\s]?\d{1,9}$
          type: string
        surname:
          description: User surname
          type: string
        roles:
          description: User roles
          uniqueItems: true
          type: array
          items:
            type: string
...
    OrderDTO:
      description: Order information
      required:
        - countOfItems
        - date
        - id
        - orderStatus
      type: object
      properties:
        id:
          description: Order ID
          minimum: 1
          type: integer
          format: int64
        date:
          description: Date of order creation
          type: string
          format: date-time
        orderStatus:
          description: Order status
          type: string
          enum:
            - PENDING
            - PROCESSING
            - SHIPPED
            - DELIVERED
            - CANCELLED
        countOfItems:
          description: Count of different products in an order
          minimum: 0
          type: integer
          format: int64
    OrdersDTO:
      description: Contains a list of orders and the total number of orders in the database
      required:
        - count
        - orderList
      type: object
      properties:
        orderList:
          description: Order list
          type: array
          items:
            $ref: '#/components/schemas/OrderDTO'
        count:
          description: Total number of orders in the database
          minimum: 0
          type: integer
          format: int64
...
    OrderItemProductDTO:
      description: Information about the product that is included in the order
      required:
        - count
        - imageLink
        - name
        - price
        - productId
      type: object
      properties:
        productId:
          description: Target product ID
          minimum: 1
          type: integer
          format: int64
        count:
          description: Product count
          minimum: 0
          type: integer
          format: int64
        price:
          description: Product price
          minimum: 0.01
          exclusiveMinimum: false
          type: number
          format: double
        name:
          description: Product name
          pattern: ^[A-zА-я0-9]{1,40}$
          type: string
        imageLink:
          description: Product image link
          type: string
    OrderItemsDTO:
      required:
        - orderItemProductDTOList
        - totalCountOfItems
      type: object
      properties:
        orderItemProductDTOList:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemProductDTO'
        totalCountOfItems:
          minimum: 0
          type: integer
          format: int64
...
    RequestDataDetailsDTO:
      description: Request details like page num and number of elements on the page
      required:
        - countPerPage
        - pageNum
      type: object
      properties:
        countPerPage:
          default: 10
          minimum: 10
          type: integer
          format: int32
        pageNum:
          default: 1
          minimum: 1
          type: integer
          format: int32
...
    UsersDTO:
      required:
        - userDTOList
        - count
      type: object
      properties:
        userDTOList:
          type: array
          items:
            $ref: "#/components/schemas/UserDTO"
        count:
          type: integer
          format: int64
  securitySchemes:
    Bearer-Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT